班级管理与学生管理独立实现

+++ web-code/web-ai-project02/tlias-web-management/src/main/java/com/itheima/service/impl/DeptServiceImpl.java
@@ -1,5 +1,6 @@
 package com.itheima.service.impl;
 
+import com.itheima.exception.GlobalExceptionHandler;
 import com.itheima.mapper.DeptMapper;
 import com.itheima.pojo.Dept;
 import com.itheima.service.DeptService;
@@ -22,7 +23,12 @@ public class DeptServiceImpl implements DeptService {
 
     @Override
     public void deleteById(Integer id){
-        deptMapper.deleteById(id);
+        //如果部门下有员工，则不允许删除该部门，并给前端提示错误信息：对不起，当前部门下有员工，不能直接删除！
+        if (deptMapper.getEmpCountByDeptId(id) > 0){
+            throw new GlobalExceptionHandler.ClazzHasStudentException("对不起，当前部门下有员工，不能直接删除！");
+        }else{
+            deptMapper.deleteById(id);
+        }
     }
 
     @Override
